backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20

# Data Processing
pandas==2.2.3
numpy==2.2.2
pyarrow==19.0.0

# Machine Learning
scikit-learn==1.6.1
joblib==1.4.2
scipy==1.15.1

# API & Validation
pydantic==2.10.6
requests==2.32.3

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Caching & Rate Limiting
redis==5.0.1
slowapi==0.1.9

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.40.0

# Utilities
jinja2==3.1.5
pyyaml==6.0.1

# Development (optional, comment out for production)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
# black==23.12.1
# flake8==6.1.0
